requirements.txt
streamlit==1.48.0
sentence-transformers==5.1.0
psycopg2-binary==2.9.10
tqdm==4.67.1
pgvector==0.3.6
requests==2.32.4
beautifulsoup4==4.13.4
tabulate==0.9.0
unidecode==1.4.0
openai==1.93.3